% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNNregR-package.R
\docType{package}
\name{CNNregR-package}
\alias{CNNregR-package}
\alias{CNNregR}
\title{CNNregR: R Interface for CNNreg Cell Type Deconvolution}
\description{
CNNregR provides R functions for preprocessing scRNA-seq and bulk RNA-seq
data for CNNreg deconvolution analysis using edgeR TMM normalization and
k-means clustering. It includes:

- Marker gene selection with edgeR TMM normalization and discrimination scoring
- Optional gene biotype filtering via BioMart for multiple organisms
- K-means clustering-based reference generation from scRNA-seq
- Data formatting for CNNreg compatibility
- Wrapper functions to run CNNreg from R
}
\section{Main Functions}{


**High-level workflow:**
\itemize{
  \item \code{\link{deconvolve}}: Complete preprocessing + deconvolution pipeline
}

**Preprocessing:**
\itemize{
  \item \code{\link{select_genes}}: Select discriminative genes with edgeR TMM
  \item \code{\link{preprocess_bulk}}: Format bulk RNA-seq data
  \item \code{\link{preprocess_reference}}: Generate reference CSE profiles via k-means
}

**Deconvolution:**
\itemize{
  \item \code{\link{run_cnnreg}}: Call CNNreg CLI from R
}

**Utilities:**
\itemize{
  \item \code{\link{check_cnnreg_install}}: Verify CNNreg installation
  \item \code{\link{validate_input_data}}: Check data format
  \item \code{\link{format_seurat_object}}: Convert Seurat to CNNregR format
  \item \code{\link{get_gene_annotation}}: Query BioMart for gene annotations
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mwang159/CNNregR}
  \item Report bugs at \url{https://github.com/mwang159/CNNregR/issues}
}

}
\author{
\strong{Maintainer}: Xue Wang \email{wang.xue@mayo.edu}

Authors:
\itemize{
  \item Yuanhang Liu \email{liu.yuanhang@mayo.edu}
}

}
\keyword{internal}
