% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Get Gene Annotation from BioMart}
\usage{
get_gene_annotation(
  genes,
  organism = "hsapiens",
  attributes = c("hgnc_symbol", "gene_biotype", "chromosome_name"),
  host = "https://www.ensembl.org"
)
}
\arguments{
\item{genes}{Character vector of gene symbols}

\item{organism}{Organism name for BioMart. Options:
- "hsapiens" (human, default)
- "mmusculus" (mouse)
Or full dataset name like "hsapiens_gene_ensembl"}

\item{attributes}{Attributes to retrieve (default: gene symbol, biotype, chromosome)}

\item{host}{BioMart host (default: "https://www.ensembl.org")}
}
\value{
Data.frame with gene annotations
}
\description{
Retrieves gene annotations (biotype, chromosome) from Ensembl BioMart
for specified organism
}
\examples{
\dontrun{
# Human genes
anno <- get_gene_annotation(
  genes = c("TP53", "GAPDH", "ACTB"),
  organism = "hsapiens"
)

# Mouse genes
anno_mouse <- get_gene_annotation(
  genes = c("Tp53", "Gapdh"),
  organism = "mmusculus"
)
}
}
