% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_cnnreg_proportions}
\alias{read_cnnreg_proportions}
\title{Read CNNreg Output Proportions}
\usage{
read_cnnreg_proportions(file, rename_samples = NULL)
}
\arguments{
\item{file}{Path to Prop_predicted_*.csv file}

\item{rename_samples}{Optional named vector to rename samples}
}
\value{
Data.frame with cell proportions
}
\description{
Helper function to read cell proportion predictions from CNNreg output
}
\examples{
\dontrun{
props <- read_cnnreg_proportions("output/Prop_predicted_GBM.csv")
}
}
