% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate Input Data Format}
\usage{
validate_input_data(bulk_counts, sc_counts_list)
}
\arguments{
\item{bulk_counts}{Bulk RNA-seq count matrix}

\item{sc_counts_list}{List of scRNA-seq count matrices}
}
\value{
List with validation results and warnings
}
\description{
Checks that input data meets CNNreg requirements
}
